requirements.txt
fastapi==0.119.0
uvicorn==0.22.0
biopython==1.85
certifi==2025.10.5
google-genai==1.45.0
